auth middleware unit tests

+++ src/middleware/__tests__/auth.tests.ts
@@ -3,14 +3,22 @@ import { GenerateSignature } from "../../utils/auth-utils";
 import { UserInstance } from '../../model/userModel'
 import {v4 as UUID} from 'uuid'
 import app from "../../server"
+import { auth } from "../auth/auth";
+import { Response } from "express-serve-static-core";
+import { JwtPayload } from "jsonwebtoken";
+import { NextFunction } from "express";
 
 const email = "SMOOZE@gmail.com"
 const id1=UUID()
 const id2=UUID()
+const id3=UUID()
+let request:supertest.SuperTest<supertest.Test> = supertest(app)
+let token :string|null;
+let adminToken :string|null;
+
+
 beforeAll(async()=>{
     try{
-
-
     await UserInstance.bulkCreate(
         [{
         id:id1,
@@ -37,6 +45,20 @@ beforeAll(async()=>{
         country:"Nigeria",
         gender:"male",
         verified:true,
+    },
+    {
+        id:id3,
+        password:"1234abcd",
+        salt:"hello",
+        userName:"Smoozeadmin",
+        is_premium:false,
+        firstName:"SMOOZE",
+        lastName:"SMOOZE",
+        email:"smoozeadmintest@gmail.com",
+        country:"Nigeria",
+        gender:"male",
+        verified:true,
+        role:"admin"
     }
 ])
 }
@@ -46,11 +68,9 @@ catch(err:any){
 })
 
 afterAll(async()=>{
-    await UserInstance.destroy({ where: { email:[email, "unverifiedUser@gmail.com"] } });
+    await UserInstance.destroy({ where: { email:[email, "unverifiedUser@gmail.com","smoozeadmintest@gmail.com"] } });
 })
 
-let request:supertest.SuperTest<supertest.Test> = supertest(app)
-let token :string|null;
 describe("auth middleware integration tests",()=>{
     const exec = () => {
         return request
@@ -87,11 +107,15 @@ describe("auth middleware integration tests",()=>{
         token = ""
         const response = await exec();
         expect(response.status).toBe(400)
+        expect(response.body.error).toBe("jwt must be provided")
+
     })
 
     it("should return 400 if token is correct but user is invalid", async()=>{
         const response = await exec();
         expect(response.status).toBe(400)
+        expect(response.body.error).toBe("jwt must be provided")
+
     })
 
     it("should return 400 if token is correct and user is valid but user is not verified", async()=>{
@@ -109,8 +133,171 @@ describe("auth middleware integration tests",()=>{
     })
 })
 
+describe("admin auth middleware integration tests",()=>{
+    const exec = () => {
+        return request
+        .patch('/api/admin/update')
+        .set("Authorization", `Bearer ${adminToken}`)
+        .trustLocalhost(true)
+        .send({ 
+            firstName:"SMOOZE",
+            lastName:"SMOOZE",
+            email:"SMOOZE@gmail.com",
+            country:"Nigeria",
+            gender:"male"
+        })
+    }
+
+    const execNoToken = () => {
+        return request
+        .patch('/api/admin/update')
+        .trustLocalhost(true)
+        .send({ 
+            firstName:"SMOOZE",
+            lastName:"SMOOZE",
+            email:"SMOOZE@gmail.com",
+            country:"Nigeria",
+            gender:"male"
+        })
+    }
+    it("should return 400 if token is not specified", async()=>{
+        const response = await execNoToken();
+        expect(response.status).toBe(400)
+    })
+
+    it("should return 400 if token is incorrect", async()=>{
+        token = ""
+        const response = await exec();
+        expect(response.status).toBe(400)
+        expect(response.body.error).toBe("jwt must be provided")
+
+    })
+
+    it("should return 400 if token is correct but user is invalid", async()=>{
+        const response = await exec();
+        expect(response.status).toBe(400)
+        expect(response.body.error).toBe("jwt must be provided")
+
+    })
+
+    it("should return 400 if token is correct and user is valid but user is not an admin", async()=>{
+        token = await GenerateSignature({email,id:id2, verified:true, isLoggedIn: true});
+        const response = await exec();
+        expect(response.status).toBe(400)
+        expect(response.body.error).toBe("You are not an admin")
+    })
+
+    it("should return 200 if token is correct and user is an admin", async()=>{
+        token = await GenerateSignature({email:'smoozeadmintest@gmail.com',id:id3,verified:true,isLoggedIn: true});
+        const response = await exec();
+        expect(response.status).toBe(200)
+        expect(response.body.message).toBe("You have successfully updated the user profile")
+    })
+})
+
 describe("auth middleware unit tests",()=>{
-    it("should generate req.user", ()=>{
+
+    let req: Partial<JwtPayload>;
+    let res: Partial<Response>;
+    let next: NextFunction = jest.fn();
+  
+    beforeEach(() => {
+      req = {};
+      res = {
+        json: jest.fn(),
+      };
+    });
+    it("should call next function with no arguments if token is correct and verified", async()=>{
+        token = await GenerateSignature({email,id:id2,verified:true,isLoggedIn: true});
+        console.log(token)
+        req = {
+            headers:{
+                Authorization: `Bearer ${token}`
+            }
+        }
+        res = {}
+        next = jest.fn()
+
+        auth(req as JwtPayload, res as Response, next)
+        expect(next).toBeCalledTimes(1);
+        expect(next).toBeCalledWith();
+    })
+    it("should call next with error if user is not verified", async()=>{
+        token = await GenerateSignature({email:'unverified@gmail.com',id:id1,verified:true,isLoggedIn: true});
+        req = {
+            headers:{
+                Authorization: `Bearer ${token}`
+            }
+        }
+        res = {}
+        next = jest.fn()
+
+        auth(req as JwtPayload, res as Response, next)
+        expect(next).toBeCalledTimes(1);
+        expect(next).toBeCalledWith({ code: 400, message: "Account Not Activated" });
+
+    })
+
+    it("should call next with error if user does not exist", async()=>{
+        token = await GenerateSignature({email:'unverified@gmail.com',id:"1234",verified:true,isLoggedIn: true});
+        req = {
+            headers:{
+                Authorization: `Bearer ${token}`
+            }
+        }
+        res = {}
+        next = jest.fn()
+
+        auth(req as JwtPayload, res as Response, next)
+        expect(next).toBeCalledTimes(1);
+        expect(next).toBeCalledWith({ code: 400, message: "Invalide Credentials" });
+
 
     })
 })
+
+
+// describe("admin auth middleware unit tests",()=>{
+
+//     let req: Partial<JwtPayload>;
+//     let res: Partial<Response>;
+//     let next: NextFunction = jest.fn();
+  
+//     beforeEach(() => {
+//       req = {};
+//       res = {
+//         json: jest.fn(),
+//       };
+//     });
+//     it("should call next function with no arguments if user is an admin", async()=>{
+//         token = await GenerateSignature({email,id:id3,verified:true,isLoggedIn: true});
+//         console.log(token)
+//         req = {
+//             headers:{
+//                 Authorization: `Bearer ${token}`
+//             }
+//         }
+//         res = {}
+//         next = jest.fn()
+
+//         adminauth(req as JwtPayload, res as Response, next)
+//         expect(next).toBeCalledTimes(1);
+//         expect(next).toBeCalledWith();
+//     })
+//     it("should call next with error if user is not an admin", async()=>{
+//         token = await GenerateSignature({email:'unverified@gmail.com',id:id1,verified:true,isLoggedIn: true});
+//         req = {
+//             headers:{
+//                 Authorization: `Bearer ${token}`
+//             }
+//         }
+//         res = {}
+//         next = jest.fn()
+
+//         adminauth(req as JwtPayload, res as Response, next)
+//         expect(next).toBeCalledTimes(1);
+//         expect(next).toBeCalledWith({ code: 400, message: "Unauthorised" });
+
+//     })
+
+// })
